rename lib stuff
